% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_data_helpers.R
\name{filter_country}
\alias{filter_country}
\title{Filter WPP age data by country}
\usage{
filter_country(.data, country, location_column = location)
}
\arguments{
\item{.data}{A tibble containing WPP age data}

\item{country}{Name of the country}

\item{location_column}{Name of the location column, unquoted. Default \code{location}.}
}
\value{
A tibble containing WPP age data for the specified country
}
\description{
Filters WPP age data to a country
}
