% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_data_helpers.R
\name{filter_year}
\alias{filter_year}
\title{Filter WPP age data by year}
\usage{
filter_year(.data, target_year, location_column = year)
}
\arguments{
\item{.data}{A tibble containing WPP age data}

\item{target_year}{Numeric year that you want to filter on}

\item{location_column}{Name of the location column, unquoted. Default \code{year}.}
}
\value{
A tibble containing WPP age data for the specified country
}
\description{
Filters WPP age data to a year
}
