% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_data_helpers.R
\name{get_age_data}
\alias{get_age_data}
\title{Age data from the WPP API}
\usage{
get_age_data(
  location_id,
  age_bracket,
  by_sex,
  start_year = 1950,
  end_year = 2100,
  .progress = TRUE
)
}
\arguments{
\item{location_id}{tibble containing the location id of interest. Typically output from \link{get_id}.}

\item{age_bracket}{integer. Either 1 or 5, indicating the size of the brackets}

\item{by_sex}{boolean TRUE to segregate results by sex}

\item{start_year}{integer Start year for data. Default 1950 (as early as available)}

\item{end_year}{integer End year for data. Default 2100 (as late as available)}

\item{.progress}{boolean Whether to show progress bar. Default TRUE.}
}
\value{
tibble containing location_id, age information, sex, year, in a slightly
cleaner format than the raw API call.
}
\description{
Gets age data in either 1- or 5- year age brackets, potentially segregated by sex
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aus_id <- get_id("Australia", type = "locations", search = FALSE, .progress = FALSE)
get_age_data(aus_id$id, age_bracket = 5, start_year = 2023, end_year = 2023, by_sex = TRUE)
\dontshow{\}) # examplesIf}
}
