% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_levels.R
\name{get_base_levels}
\alias{get_base_levels}
\title{Get parameter inputs for population data API}
\usage{
get_base_levels(target)
}
\arguments{
\item{target}{String, one of "Indicators" or "locations"}
}
\value{
\code{tibble} containing a wide range of information about the requested endpoint
}
\description{
Gets information about indicators (i.e. dataset) and locations for further
querying from the WPP API
}
\examples{
get_base_levels("Indicators")
get_base_levels("locations")

}
