% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_levels.R
\name{get_id}
\alias{get_id}
\title{IDs of indicators}
\usage{
get_id(name, type, search = TRUE)
}
\arguments{
\item{name}{Name of the variable to search for}

\item{type}{One of "locations" or "Indicators", which dataset to search}

\item{search}{Whether to perform simple search/matching. Defaults to TRUE}
}
\value{
\code{tibble} containing only the matching records. Column \code{id} is the numeric
id which will be needed downstream
}
\description{
Get numeric IDs of indicators and locations for the WPP API
}
\examples{
get_id("Australia", typ = "locations")

}
