% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_wpp_api_key}
\alias{set_wpp_api_key}
\title{Set the WPP API key for this session}
\usage{
set_wpp_api_key(key)
}
\arguments{
\item{key}{The WPP API key to use for the rest of the session}
}
\value{
The result from \code{\link[base:Sys.setenv]{base::Sys.setenv()}}.
}
\description{
Set the WPP API key for this session
}
