% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpp_age_fiveyr}
\alias{wpp_age_fiveyr}
\title{Five year age data, by country, and year}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5040 rows and 6 columns.
}
\usage{
wpp_age_fiveyr
}
\description{
\itemize{
\item locationId Numeric location for the country
\item location Country
\item year Year of data, in this case 2023
\item ageStart Lower bound of age
\item ageEnd Upper bound of age
\item population Number of individuals
}
}
\keyword{datasets}
