% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpp_sex_fiveyr}
\alias{wpp_sex_fiveyr}
\title{Five year age data, by country and sex}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 720 rows and 5 columns.
}
\usage{
wpp_sex_fiveyr
}
\description{
\itemize{
\item locationId Numeric location for the country
\item location Country
\item year Year of data, in this case 2023
\item sexId Factor for the sex of an individual. Options are: Male, Female, Both Sexes.
Note Both Sexes is the sum of Male and Female.
\item population Number of individuals
}
}
\keyword{datasets}
